requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
joblib==1.4.2
pandas==2.2.3
numpy==2.1.3
xgboost==3.2.0
pydantic==2.7.0
python-dotenv==1.0.0